Fix selling more than 12 items at once
Added a coroutine for selling more than 12 items and also a OnUpdate frame to handle the continuous calls to the function.

OnUpdate is child of Button which will make the OnUpdate stop whenever the Button is hidden (not at a merchant). Will also hide itself when it's done selling all items.

Also added a "StaticPopup1Button1Click()" after using the container item of an item in case the item triggers the "Selling this item will make it non-tradable" popup that would cause a race condition and crash the game.